resolve toast issue

+++ ecommerce/src/context/cart.js
@@ -8,25 +8,21 @@ import { API } from "../API/endpoint";
 const CartContext = createContext();
 
 const CartProvider = ({ children }) => {
-
     const [cart, setCart] = useState([])
-
     const [auth] = useAuth()
-
     const userId = auth?.user?.userid
 
     const addToCart = async (productId) => {
         try {
-
+            if (!userId)
+                return;
             const response = await axios.post(`${API}/cart/cartAdd/${userId}`, {
                 productId,
                 quantity: 1, // Assuming you're adding one quantity at a time
             });
-
             if (response.data.success) {
                 const updatedCart = [...cart, response.data?.cart];
                 toast.success("Product add into your cart")
-
                 setCart(updatedCart);
             }
         } catch (error) {
@@ -41,28 +37,26 @@ const CartProvider = ({ children }) => {
     };
 
     const fetchCart = async () => {
+        if (!userId) return;
         try {
             const response = await axios.get(`${API}/cart/getCart/${userId}`);
             if (response.data?.cart) {
                 setCart(response.data?.cart?.items);
-                // Assuming cart.items contains the items array
             }
         } catch (error) {
             console.error('Error fetching cart:', error);
 
         }
     };
-
-
     useEffect(() => {
         if (auth?.user) {
             fetchCart();
+        } else {
+            setCart([]);
         }
     }, [auth?.user])
-
-
-
     const removeFromCart = async (productId) => {
+        if (!userId) return;
         try {
             const response = await axios.delete(`${API}/cart/cartDelete/${userId}/${productId}`);
             if (response.data.success) {
@@ -75,7 +69,6 @@ const CartProvider = ({ children }) => {
             // Handle error, display error message, etc.
         }
     };
-
 
     return <CartContext.Provider value={[cart, addToCart, removeFromCart, fetchCart]}>
         {/* //now we can use auth and set auth anywhere */}
